requirements.txt
streamlit
tensorflow-cpu==2.16.*
keras==3.3.*
torch
torchvision
timm
pillow
opencv-python
streamlit-webrtc
av
